+++ snmp/tests/test_e2e_snmp_listener.py
@@ -40,8 +40,13 @@ def test_e2e_snmp_listener(dd_agent_check, container_ip, autodiscovery_ready):
     """
     snmp_device = _build_device_ip(container_ip)
     subnet_prefix = ".".join(container_ip.split('.')[:3])
-    aggregator = dd_agent_check(
-        {'init_config': {}, 'instances': []}, rate=True, discovery_min_instances=5, discovery_timeout=10
+
+    aggregator = common.dd_agent_check_wrapper(
+        dd_agent_check,
+        {'init_config': {}, 'instances': []},
+        rate=True,
+        discovery_min_instances=5,
+        discovery_timeout=10,
     )
 
     # === network profile ===
